+++ ios/Classes/FlutterKushkiPlugin.h
@@ -1,4 +1,0 @@
-#import <Flutter/Flutter.h>
-
-@interface FlutterKushkiPlugin : NSObject<FlutterPlugin>
-@end