+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.3
+
+* Changed this project from a plugin to a package
+* Removed dependency of Java and iOS Kushki packages
+* Requests are done directly to Kushki's API
+
 ## 0.0.2
 
 * Completed get card token